fix: Use the favicon provided by the miniflux server
